Initial Explorer/Social implementation
This commit migrates all of the Explorer and Social functionality from the `lotus-website` repository.

We also integrate the Wallet into the Explorer/Social, e.g. addresses in your Contacts list are displayed as the contact name in transaction outputs.

This is just a preliminary implementation and will be expanded in future commits.