Implement Pay-To-Taproot, add Bitcore SDK plugin
This commit is the first Pay-To-Taproot (P2TR) implementation. It fully integrates P2TR into the wallet (address/script/sig/etc., Chronik, UI display). P2TR is also configured as the default address type for the wallet, with optional fallback to P2PKH.

This commit also fixes the Bitcore SDK implementation by centralizing the modules to a Nuxt plugin that is automatically loaded at app launch. This way, the SDK is fully available everywhere in the entire application, rather than using hacky conditionals and such.